requirements.txt
joblib==1.4.2
matplotlib==3.7.5
numpy==1.24.3
pandas==2.0.3
Pillow==11.1.0
Requests==2.32.3
scikit_learn==1.3.2
torch==2.4.1
torchvision==0.19.1
umap_learn==0.5.7
utils==1.0.2
